// src/api/api.types.ts
export namespace NQuakeAPI {
    export interface ILeaderboard {
        boardType: string;
        entries: ILeaderboardItem[];
        totalEntries: number;
    }

    export interface ILeaderboardItem {
        userName: string;
        eloRating: number;
        profileIconId: string;
        namePlateId: string;
    }
// ...
}

export namespace NApp {
    export interface IStreamer {
        elo?: number;
        name?: string;
        iconId?: string;
        namePlateId?: string;
    }
}